1.修改字段类型

+++ nascent-bridge/src/main/java/com/me/nascent/modules/member/entity/Member.java
@@ -101,10 +101,10 @@ public class Member {
     private Integer sgRecruitGuideId;
 
     @TableField("sgRecruitShopId")
-    private Integer sgRecruitShopId;
+    private Long sgRecruitShopId;
 
     @TableField("sgShopId")
-    private Integer sgShopId;
+    private Long sgShopId;
 
     @TableField("source")
     private String source;
